add namespaced consumer group id on flink engine
