+++ engine/flink/kafka-components-utils/src/main/scala/pl/touk/nussknacker/engine/kafka/generic/FlinkKafkaDelayedSourceImplFactory.scala
@@ -4,6 +4,7 @@ import org.apache.flink.api.common.eventtime.SerializableTimestampAssigner
 import org.apache.flink.streaming.api.functions.source.SourceFunction
 import org.apache.kafka.clients.consumer.ConsumerRecord
 import pl.touk.nussknacker.engine.api.context.transformation.NodeDependencyValue
+import pl.touk.nussknacker.engine.api.namespaces.NamingStrategy
 import pl.touk.nussknacker.engine.api.process.{ContextInitializer, Source}
 import pl.touk.nussknacker.engine.flink.api.process.FlinkCustomNodeContext
 import pl.touk.nussknacker.engine.flink.api.timestampwatermark.{
@@ -51,7 +52,8 @@ class FlinkKafkaDelayedSourceImplFactory[K, V](
       deserializationSchema: KafkaDeserializationSchema[ConsumerRecord[K, V]],
       formatter: RecordFormatter,
       contextInitializer: ContextInitializer[ConsumerRecord[K, V]],
-      testParametersInfo: KafkaTestParametersInfo
+      testParametersInfo: KafkaTestParametersInfo,
+      namingStrategy: NamingStrategy
   ): Source = {
     extractDelayInMillis(params) match {
       case millis if millis > 0 =>
@@ -68,7 +70,8 @@ class FlinkKafkaDelayedSourceImplFactory[K, V](
           formatter,
           contextInitializer,
           testParametersInfo,
-          millis
+          millis,
+          namingStrategy
         )
       case _ =>
         super.createSource(
@@ -80,7 +83,8 @@ class FlinkKafkaDelayedSourceImplFactory[K, V](
           deserializationSchema,
           formatter,
           contextInitializer,
-          testParametersInfo
+          testParametersInfo,
+          namingStrategy
         )
     }
   }
@@ -93,7 +97,8 @@ class FlinkKafkaDelayedSourceImplFactory[K, V](
       formatter: RecordFormatter,
       contextInitializer: ContextInitializer[ConsumerRecord[K, V]],
       testParametersInfo: KafkaTestParametersInfo,
-      delay: Long
+      delay: Long,
+      namingStrategy: NamingStrategy
   ): FlinkKafkaSource[ConsumerRecord[K, V]] = {
     val delayCalculator = new FixedDelayCalculator(delay)
     createDelayedKafkaSource(
@@ -104,7 +109,8 @@ class FlinkKafkaDelayedSourceImplFactory[K, V](
       formatter,
       contextInitializer,
       testParametersInfo,
-      delayCalculator
+      delayCalculator,
+      namingStrategy
     )
   }
 
@@ -116,7 +122,8 @@ class FlinkKafkaDelayedSourceImplFactory[K, V](
       formatter: RecordFormatter,
       contextInitializer: ContextInitializer[ConsumerRecord[K, V]],
       testParametersInfo: KafkaTestParametersInfo,
-      delayCalculator: DelayCalculator
+      delayCalculator: DelayCalculator,
+      namingStrategy: NamingStrategy
   ): FlinkKafkaSource[ConsumerRecord[K, V]] = {
     new FlinkConsumerRecordBasedKafkaSource[K, V](
       preparedTopics,
@@ -125,7 +132,8 @@ class FlinkKafkaDelayedSourceImplFactory[K, V](
       timestampAssigner,
       formatter,
       contextInitializer,
-      testParametersInfo
+      testParametersInfo,
+      namingStrategy
     ) {
 
       override protected def createFlinkSource(
