+++ engine/flink/kafka-components-utils/src/main/scala/pl/touk/nussknacker/engine/kafka/source/flink/FlinkKafkaSourceImplFactory.scala
@@ -2,6 +2,7 @@ package pl.touk.nussknacker.engine.kafka.source.flink
 
 import org.apache.kafka.clients.consumer.ConsumerRecord
 import pl.touk.nussknacker.engine.api.context.transformation.NodeDependencyValue
+import pl.touk.nussknacker.engine.api.namespaces.NamingStrategy
 import pl.touk.nussknacker.engine.api.process.{ContextInitializer, Source}
 import pl.touk.nussknacker.engine.flink.api.timestampwatermark.TimestampWatermarkHandler
 import pl.touk.nussknacker.engine.kafka.serialization.KafkaDeserializationSchema
@@ -22,7 +23,8 @@ class FlinkKafkaSourceImplFactory[K, V](
       deserializationSchema: KafkaDeserializationSchema[ConsumerRecord[K, V]],
       formatter: RecordFormatter,
       contextInitializer: ContextInitializer[ConsumerRecord[K, V]],
-      testParametersInfo: KafkaTestParametersInfo
+      testParametersInfo: KafkaTestParametersInfo,
+      namingStrategy: NamingStrategy
   ): Source =
     new FlinkConsumerRecordBasedKafkaSource[K, V](
       preparedTopics,
@@ -31,7 +33,8 @@ class FlinkKafkaSourceImplFactory[K, V](
       timestampAssigner,
       formatter,
       contextInitializer,
-      testParametersInfo
+      testParametersInfo,
+      namingStrategy
     )
 
 }
