+++ utils/kafka-components-utils/src/main/scala/pl/touk/nussknacker/engine/kafka/source/KafkaSourceFactory.scala
@@ -7,6 +7,7 @@ import pl.touk.nussknacker.engine.api.component.UnboundedStreamComponent
 import pl.touk.nussknacker.engine.api.context.transformation._
 import pl.touk.nussknacker.engine.api.context.{ProcessCompilationError, ValidationContext}
 import pl.touk.nussknacker.engine.api.definition._
+import pl.touk.nussknacker.engine.api.namespaces.NamingStrategy
 import pl.touk.nussknacker.engine.api.process._
 import pl.touk.nussknacker.engine.api.test.TestRecord
 import pl.touk.nussknacker.engine.api.typed.typing.{Typed, TypingResult, Unknown}
@@ -168,7 +169,8 @@ class KafkaSourceFactory[K: ClassTag, V: ClassTag](
       deserializationSchema,
       formatter,
       contextInitializer,
-      KafkaTestParametersInfo.empty
+      KafkaTestParametersInfo.empty,
+      modelDependencies.namingStrategy
     )
   }
 
@@ -219,7 +221,8 @@ object KafkaSourceFactory {
         deserializationSchema: KafkaDeserializationSchema[ConsumerRecord[K, V]],
         formatter: RecordFormatter,
         contextInitializer: ContextInitializer[ConsumerRecord[K, V]],
-        testParametersInfo: KafkaTestParametersInfo
+        testParametersInfo: KafkaTestParametersInfo,
+        namingStrategy: NamingStrategy
     ): Source
 
   }
