+++ utils/schemed-kafka-components-utils/src/main/scala/pl/touk/nussknacker/engine/schemedkafka/source/UniversalKafkaSourceFactory.scala
@@ -180,7 +180,8 @@ class UniversalKafkaSourceFactory(
       deserializationSchema,
       recordFormatter,
       kafkaContextInitializer,
-      prepareKafkaTestParametersInfo(valueSchemaUsedInRuntime, preparedTopic.original)
+      prepareKafkaTestParametersInfo(valueSchemaUsedInRuntime, preparedTopic.original),
+      modelDependencies.namingStrategy
     )
   }
